Cast size_t value when using it with printf() in a test.
Use %lu with a cast to avoid warnings in both 32 and 64 bit builds.

Closes #16720.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@78261 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ tests/interactive/output.cpp
@@ -139,7 +139,7 @@ void InteractiveOutputTestCase::TestMimeEnum()
 
     size_t count = wxTheMimeTypesManager->EnumAllFileTypes(mimetypes);
 
-    wxPrintf(wxT("*** All %u known filetypes: ***\n"), count);
+    wxPrintf(wxT("*** All %lu known filetypes: ***\n"), static_cast<unsigned long>(count));
 
     wxArrayString exts;
     wxString desc;
